+++ src/shell/layershell.rs
@@ -273,7 +273,7 @@ impl State for LayerSurfaceInner {
             .unwrap();
     }
 
-    fn pointer_axis(&mut self, horizontal: u64, vertical: u64, time: u32) {
+    fn pointer_axis(&mut self, horizontal: f64, vertical: f64, time: u32) {
         self.display_tx
             .send(Event::PointerAxis {
                 horizontal,
